backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
PyMySQL
bcrypt
python-dateutil
SQLAlchemy
Werkzeug
cryptography
